+++ R/barplot.R
@@ -71,6 +71,7 @@ ui_g_barplot <- function(id,
                          summary_funs,
                          pre_output,
                          post_output) {
+  checkmate::assert_class(data, "teal_data")
   ns <- NS(id)
   teal.widgets::standard_layout(
     encoding = div(
@@ -120,7 +121,8 @@ srv_g_barplot <- function(id,
                           summary_funs) {
   with_reporter <- !missing(reporter) && inherits(reporter, "Reporter")
   assert_class(filter_panel_api, "FilterPanelAPI")
-  assert_class(data, "tdata")
+  checkmate::assert_class(data, "reactive")
+  checkmate::assert_class(shiny::isolate(data()), "teal_data")
   moduleServer(id, function(input, output, session) {
     experiment <- experimentSpecServer(
       "experiment",