+++ R/forestplot.R
@@ -98,6 +98,7 @@ ui_g_forest_tte <- function(id,
                             summary_funs,
                             pre_output,
                             post_output) {
+  checkmate::assert_class(data, "teal_data")
   ns <- NS(id)
   teal.widgets::standard_layout(
     encoding = div(
@@ -143,7 +144,8 @@ srv_g_forest_tte <- function(id,
                              plot_width) {
   with_reporter <- !missing(reporter) && inherits(reporter, "Reporter")
   assert_class(filter_panel_api, "FilterPanelAPI")
-  assert_class(data, "tdata")
+  checkmate::assert_class(data, "reactive")
+  checkmate::assert_class(shiny::isolate(data()), "teal_data")
 
   moduleServer(id, function(input, output, session) {
     experiment <- experimentSpecServer(