+++ R/scatterplot.R
@@ -71,6 +71,7 @@ ui_g_scatterplot <- function(id,
                              summary_funs,
                              pre_output,
                              post_output) {
+  checkmate::assert_class(data, "teal_data")
   ns <- NS(id)
 
   smooth_method_choices <- c(
@@ -119,7 +120,8 @@ srv_g_scatterplot <- function(id,
                               summary_funs) {
   with_reporter <- !missing(reporter) && inherits(reporter, "Reporter")
   assert_class(filter_panel_api, "FilterPanelAPI")
-  assert_class(data, "tdata")
+  checkmate::assert_class(data, "reactive")
+  checkmate::assert_class(shiny::isolate(data()), "teal_data")
 
   moduleServer(id, function(input, output, session) {
     experiment <- experimentSpecServer(